requirements.txt
openai>=1.20.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0 